Rename source files and variables
