+++ src/monolithic/gst/inference_elements/base/gva_base_inference.cpp
@@ -16,7 +16,7 @@
 #include "dlstreamer/gst/frame.h"
 #include "gstgvaclassify.h"
 #include "inference_backend/buffer_mapper.h"
-#include "inference_impl.h"
+#include "inference_coordinator.h"
 #include "model_proc_provider.h"
 
 #include "gva_base_inference_priv.hpp"
