+++ src/monolithic/gst/inference_elements/base/inference_coordinator.h
