+++ src/monolithic/gst/inference_elements/common/post_processor/post_processor_c.cpp
@@ -10,15 +10,15 @@
 #include "utils.h"
 #include <inference_backend/logger.h>
 
-post_processing::PostProcessor *createPostProcessor(InferenceCoordinator *inference_impl,
+post_processing::PostProcessor *createPostProcessor(InferenceCoordinator *coordinator,
                                                     GvaBaseInference *base_inference) {
-    if (inference_impl == nullptr) {
+    if (coordinator == nullptr) {
         GVA_WARNING("InferenceCoordinator is null. Creating of inference post processor is imposible");
         return nullptr;
     }
     PostProcessor *post_processor = nullptr;
     try {
-        post_processor = new post_processing::PostProcessor(inference_impl, base_inference);
+        post_processor = new post_processing::PostProcessor(coordinator, base_inference);
     } catch (const std::exception &e) {
         GVA_ERROR("Couldn't create post-processor: %s", Utils::createNestedErrorMsg(e).c_str());
     }
